// euc-lab-portal/src/app/models/PolicyNatRule.ts
export class PolicyNatRule {
    create_time: any;
    create_user: string;
    last_modified_time: any;
    last_modified_user: string;
    protection: string;
    revision: number;
    action: string;
    display_name: string;
    enabled: boolean;
    firewall_match: string;
    id: string;
    parent_path: string;
    path: string;
    relative_path: string;
    resource_type: string;
    scope: string[];
    service: string;
    source_network: string;
    translated_network: string;
    translated_ports: string;
    marked_for_delete: boolean;
    overridden: string;
    logging: boolean;
    tags: string;
}